# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HMRC API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/finance": {
            "get": {
                "tags": [
                    "Finance"
                ],
                "summary": "Get list transaction",
                "description": "mengambil data riwayat transaksi (masuk dan keluar)",
                "operationId": "get/finance",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit value is number by default limit is 10. ex : ?limit=100",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    },
                    {
                        "name": "bulan",
                        "in": "query",
                        "description": "Untuk filter transaksi perbulan",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "01"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return json array data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Finance"
                ],
                "summary": "Store transaction data",
                "description": "input data keuangan (pemasukan/pengeluaran)",
                "operationId": "finance/store",
                "requestBody": {
                    "description": "form data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "keterangan",
                                    "tanggal"
                                ],
                                "properties": {
                                    "keterangan": {
                                        "type": "string"
                                    },
                                    "masuk": {
                                        "type": "number"
                                    },
                                    "keluar": {
                                        "type": "number"
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return json array data"
                    }
                }
            }
        },
        "/api/finance/{id}": {
            "delete": {
                "tags": [
                    "Finance"
                ],
                "summary": "Delete transaction data",
                "description": "hapus data transaksi keuangan",
                "operationId": "finance/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "insert transaction id on path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return json array data"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Finance"
                ],
                "summary": "Edit transaction data",
                "description": "edit data transaksi",
                "operationId": "finance/edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "insert transaction id on path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "form data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "keterangan": {
                                        "type": "string"
                                    },
                                    "masuk": {
                                        "type": "number"
                                    },
                                    "keluar": {
                                        "type": "number"
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return json array data"
                    }
                }
            }
        },
...
    }
}